==> src/scripts/07_01_TE-DEC_cistrome.R <==
# --------------------------------- Notes--- -----------------------------------
# 
# This script is used to identify differentially expressed CAGE-peaks (DECs) that
# intersect with a TE. These CAGE-peaks are stored in a .bed-format that can be
# uploaded to the cistrome db.
#
# Output data (../tables/):
#
# 07_01_[tissue].te_decs.[direction].bed - bedfiles of all DECs that intersect
#                                          with a TE and separated by the 
#                                          direction
#
